requirements.txt
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
flask==3.0.0
future==0.18.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
pyngrok==7.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
urllib3==2.0.7
webexteamsbot==0.1.4.2
webexteamssdk==1.6.1
werkzeug==3.0.0
zipp==3.17.0